sys-auth/realmd: Add new patches
Realmd didn't have dev-util/intltool as a dependency, but it actually
required it during build. Apply a patch that converts the project from
intltool to gettext in order to get rid of the dependency on the
obsolete tool. To apply the patch without conflicts, apply also
another patch that modernizes the configure.ac file.